requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
pytest>=7.0.0
beautifulsoup4>=4.9.3
requests>=2.25.1
pyyaml>=5.4.1
playwright>=1.20.0
typing-extensions>=4.0.0
pandas>=1.2.0 